file restructure added descriptive folders for components
