+++ frontend/src/components/StudentDashboard/HeaderStudentDashboard.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { FaBars, FaTimes, FaHome } from "react-icons/fa";
 import { IoIosNotifications, IoMdSettings } from "react-icons/io";
-import Logo from "../assets/images/logobetter.png";
+import Logo from "../../assets/images/logobetter.png";
 const SettingsModal = ({ isOpen, onClose }: { isOpen: boolean; onClose: () => void }) => {
   const [firstName, setFirstName] = useState('');
   const [lastName, setLastName] = useState('');
