+++ frontend/src/components/StudentDashboard/JournalingModal.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
-import { useAuth } from "../hooks/useAuth";
+import { useAuth } from "../../hooks/useAuth";
 
 interface JournalEntry {
   id: number;
