+++ frontend/src/components/StudentDashboard/StudentHelpModal.tsx
