+++ frontend/src/components/TherapistDashboard/HeaderTherapistDashboard.tsx
