+++ frontend/src/components/TherapistDashboard/TherapistHelpModal.tsx
